Refactor GoToInsn class
